requirements.txt
# Core Django
Django==5.1.6
django-allauth==65.4.1
django-database-url==1.0.3
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.10

# Authentication & Social Login
social-auth-app-django==5.4.3
social-auth-core==4.5.6
oauthlib==3.2.2
PyJWT==2.10.1
defusedxml==0.7.1
python3-openid==3.2.0
requests-oauthlib==2.0.0

# Real-time & Channels
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0

# Machine Learning and Data Processing
tensorflow>=2.12.0
scikit-learn>=1.3.0
joblib>=1.3.2
numpy>=1.24.0

# File Storage & Media
Pillow>=10.0.0

# External Services
supabase>=2.0.0
requests>=2.32.3

# Production Dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0

# Security & Utilities
cryptography>=44.0.2
certifi>=2025.1.31
urllib3>=2.3.0
